# src/api.yaml
swagger: "2.0"
info:
  description: "This is a price service."
  version: "1.0.0"
  title: "Swagger PriceFollow"
basePath: "/v2"
tags:
- name: "search"
  description: "Follow price by search"
schemes:
- "http"
paths:
  /search:
    post:
      tags:
      - "search"
      summary: "新增一个search(查找)"
      description: "search与纬雅的一个产品关联，它将跟踪友商的同产品价格"
      operationId: "addSearch"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "创建search所需要的信息"
        required: true
        schema:
          $ref: "#/definitions/addsearch"
      responses:
        200:
          description: "创建search成功，返回所有符合条件的友商"
          schema:
            type: object
            items: 
              $ref: "#/definitions/search"
    get:
      tags:
      - "search"
      summary: "查找产品列表"
      description: "通过用户输入的keywords，查找出所有的产品列表"
      operationId: "querySearch"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "keywords"
        in: "query"
        description: "用与查找的关键字，各关键字之前,分隔"
        required: true
        type: "integer"
        format: "int64"      
      responses:
        200:
          description: "正确执行"
          schema:
            type: array
            items: 
              $ref: "#/definitions/search"              
  /search/{searchId}:
    delete:
      tags:
      - "search"
      summary: "删除一个search"
      description: "通过id删除一个search"
      operationId: "deleteSearch"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "searchId"
        in: "path"
        description: "search ID"
        required: true
        type: "integer"
        format: "int64"      
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Search not found"
        405:
          description: "Validation exception" 
    get:
      tags:
      - "search"
      summary: "得到search"
      description: "通过search id得到search"
      operationId: "getSearcht"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "searchId"
        in: "path"
        description: "ID of search to return"
        required: true
        type: "integer"
        format: "int64"      
      responses:
        200:
          description: "Add search successfully"
          schema:
            type: object
            items: 
              $ref: "#/definitions/search"
  /search/{search_id}/price/{price_id}:
    delete:
      tags:
      - "search"
      summary: "删除一个price"
      description: "删除一个针对友商的price跟踪"
      operationId: "deletePrice"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "price_id"
        in: "path"
        description: "price ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "search_id"
        in: "path"
        description: "这个price所在的search"
        required: true
        type: "integer"
        format: "int64"          
      responses:
        204:
          description: "删除成功" 
  /search/{search_id}/price:          
    post:
      tags:
      - "search"
      summary: "新增价格follow"
      description: "新增一个对友商的价格跟踪"
      operationId: "addPrice"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "search_id"
        in: "path"
        description: "search id,这个price将加在这个search下面"
        required: true
        type: "integer"
        format: "int64"      
      - in: "body"
        name: "body"
        description: "加入这个price所需要的信息"
        required: true
        schema:
          $ref: "#/definitions/addPrice"
      responses:
        200:
          description: "新增成功"
          schema:
            type: object
            items: 
              $ref: "#/definitions/price"          
  /search/product/{product_id}:
    get:
      tags:
      - "search"
      summary: "通过product id返回search"
      description: "通过product id返回search, product id 为https://mstore.ppdai.com/product/1371后面的1371和https://item.m.jd.com/product/2728763.html的2728763"
      operationId: "querySearchByProductId"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "product_id"
        in: "path"
        description: "search to return"
        required: true
        type: "integer"
        format: "int64"      
      responses:
        200:
          description: "返回search"
          schema:
            type: array
            items: 
              $ref: "#/definitions/search"
  /search/min:
    get:
      tags:
      - "search"
      summary: "返回纬雅的价格最小的所有产品"
      description: "返回纬雅的价格最小的所有产品"
      operationId: "querySearchLtMin"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Query search successfully"
          schema:
            type: array
            items: 
              $ref: "#/definitions/search" 
  /search/avg:
    get:
      tags:
      - "search"
      summary: "返回纬雅的价格小于均价的所有产品"
      description: "返回纬雅的价格小于均价的所有产品"
      operationId: "querySearchLtAvg"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Query search successfully"
          schema:
            type: array
            items: 
              $ref: "#/definitions/search"               
definitions:
  search:
    type: "object"
    required:
    - "keywords"
    properties:
      id:
        type: "integer"
        description: "Search ID, if create"
      keywords:
        type: "string"
        description: "And words, if all words in content, accept"
        example: "iphone,7"
      e_keywords:
        type: "string"
        description: "Exclusive words, if these words in content, skip"
        example: "plus"
      o_keywords:
        type: "string"
        description: "Or words, if any word in contect, accept"
        example: "plus"
      description:
        type: "string"
        example: "iphone 7 plus 金色"
      min:
        type: "string"
        example: "14.7"
        description: "最小价格"
      max:
        type: "string"
        example: "14.7"
        description: "最大价格"
      avg:
        type: "string"
        example: "14.7"
        description: "平均价格"        
      prices:
        type: array
        description: "所有友商的价格列表"
        items:
          $ref: "#/definitions/price"
      target:
        $ref: "#/definitions/price"
      create_date:
        type: "string"
        description: "Create Datetime"
      update_date:
        type: "string"
        description: "Update Datetime"
  addsearch:
    type: "object"
    required:
    - "keywords"
    - "url"
    properties:
      keywords:
        type: "string"
        description: "And words, if all words in content, accept"
        example: "iphone,7"
      e_keywords:
        type: "string"
        description: "Exclusive words, if these words in content, skip"
        example: "plus"
      o_keywords:
        type: "string"
        description: "Or words, if any word in contect, accept"
        example: "plus"
      url:
        type: "string"
        example: "https://mstore.ppdai.com/product/1390"
      description:
        type: "string"
        example: "iphone 7 plus 金色"
  price:
    type: "object"
    required:
    - "search_id"
    - "url"
    properties:
      id:
        type: "integer"
        description: "Search ID, if create, please ignore"
      search_id:
        type: "string"
        description: "search id"
        example: "1"
      product_id:
        type: "string"
        description: "product id"
        example: "1"
      description:
        type: "string"
      src:
        type: "string"
        example: "pp|jd"
      price:
        type: "number"
        example: "4099"
      seller:
        type: "string"
        example: "纬雅|京东"
      url:
        type: "string"
        example: "iphone 7 plus 金色"        
      create_date:
        type: "string"
        description: "Create Datetime"
      update_date:
        type: "string"
        description: "Update Datetime"
  addPrice:
    type: "object"
    required:
    - "url"
    properties:
      url:
        type: "string"
        description: "拍拍或者京东的URL"
        example: "https://mstore.ppdai.com/product/1"
      seller:
        type: "string"
        description: "卖家"
        example: "飞牛网"